refactor: placeMapDatasource 제거
