--- android/app/src/main/java/com/daedan/festabook/data/datasource/remote/placeMap/PlaceMapDataSource.kt
@@ -1,11 +1,0 @@
-package com.daedan.festabook.data.datasource.remote.placeMap
-
-import com.daedan.festabook.data.datasource.remote.ApiResult
-import com.daedan.festabook.data.model.response.OrganizationGeographyResponse
-import com.daedan.festabook.data.model.response.PlaceGeographyResponse
-
-interface PlaceMapDataSource {
-    suspend fun fetchOrganizationGeography(): ApiResult<OrganizationGeographyResponse>
-
-    suspend fun fetchPlaceGeographies(): ApiResult<List<PlaceGeographyResponse>>
-}
