+++ android/app/src/main/java/com/daedan/festabook/data/repository/PlaceListRepositoryImpl.kt
@@ -1,7 +1,6 @@
 package com.daedan.festabook.data.repository
 
 import com.daedan.festabook.data.datasource.remote.place.PlaceDataSource
-import com.daedan.festabook.data.datasource.remote.placeMap.PlaceMapDataSource
 import com.daedan.festabook.data.model.response.toDomain
 import com.daedan.festabook.data.util.toResult
 import com.daedan.festabook.domain.model.OrganizationGeography
@@ -11,7 +10,6 @@ import com.daedan.festabook.domain.repository.PlaceListRepository
 
 class PlaceListRepositoryImpl(
     private val placeDataSource: PlaceDataSource,
-    private val placeMapDataSource: PlaceMapDataSource,
 ) : PlaceListRepository {
     override suspend fun getPlaces(): Result<List<Place>> {
         val response = placeDataSource.fetchPlaces().toResult()
@@ -19,12 +17,12 @@ class PlaceListRepositoryImpl(
     }
 
     override suspend fun getOrganizationGeography(): Result<OrganizationGeography> {
-        val response = placeMapDataSource.fetchOrganizationGeography().toResult()
+        val response = placeDataSource.fetchOrganizationGeography().toResult()
         return response.map { it.toDomain() }
     }
 
     override suspend fun getPlaceGeographies(): Result<List<PlaceGeography>> {
-        val response = placeMapDataSource.fetchPlaceGeographies().toResult()
+        val response = placeDataSource.fetchPlaceGeographies().toResult()
         return response.map { placeGeographies -> placeGeographies.map { it.toDomain() } }
     }
 }
